rename http_server to storage_server to match class name
