+++ src/service_node/storage_server.cpp
@@ -1,4 +1,4 @@
-#include "http_server.h"
+#include "storage_server.h"
 #include "pow.hpp"
 #include "utils.hpp"
